rename default placeholder components
