+++ packages/sitecore-jss-angular/src/components/date.directive.ts
@@ -9,7 +9,7 @@ import {
 } from '@angular/core';
 import { DateField } from './rendering-field';
 import { BaseFieldDirective } from './base-field.directive';
-import { DefaultEmptyFieldEditingComponent } from './default-empty-field-editing.component';
+import { DefaultEmptyFieldEditingComponent } from './default-empty-text-field-editing-placeholder.component';
 
 @Directive({
   selector: '[scDate]',
