+++ packages/sitecore-jss-angular/src/components/default-empty-text-field-editing-placeholder.component.ts
@@ -4,7 +4,7 @@ import { Component } from '@angular/core';
  * Default component that will be rendered in pages when field is empty; applies for text, richtext, date and link fields.
  */
 @Component({
-  selector: 'app-default-empty-field-editing',
+  selector: 'app-default-empty-text-field-editing-placeholder',
   template: '<span>[No text in field]</span>',
 })
 export class DefaultEmptyFieldEditingComponent {}
