+++ packages/sitecore-jss-angular/src/components/rich-text.directive.ts
@@ -11,7 +11,7 @@ import { Router } from '@angular/router';
 import { isAbsoluteUrl } from '@sitecore-jss/sitecore-jss/utils';
 import { RichTextField } from './rendering-field';
 import { BaseFieldDirective } from './base-field.directive';
-import { DefaultEmptyFieldEditingComponent } from './default-empty-field-editing.component';
+import { DefaultEmptyFieldEditingComponent } from './default-empty-text-field-editing-placeholder.component';
 
 @Directive({
   selector: '[scRichText]',
